Working version
- gotta cleanup ;_;
